requirements.txt
Django>=4.2
django-filter>=23.5